[ADF-3279] Added progress bar and better error message handling
